feat: Add byte string input option for ROMA InvocationRequests
InvocationRequest in roma.h now has a boolean to treat the input in
`input` as a byte string instead of a JSON escaped string. When
`treat_input_as_byte_str` is set to `true`, `input` is expected to be of
size 1 because all inputs can be serialized to this single byte string
as an input. This field is temporary to support byte string inputs into
ROMA. A future iteration will allow for more control of inputs and each
input type.

Bug: b/314287539
Change-Id: Idd71f73bfe9e1bd53814f36f64b3cef955562f39
GitOrigin-RevId: 14c59338a332b62842b2c950e3585d8fd783a4e0