feat: Add callback service for gRPC server
Adds functionality to recieve a serialized request proto from a UDF, send it to the host process via gRPC, where the appropriate function based on the js_function_name is invoked with the now deserialized request proto. This function would then return a serialized response proto, which is sent directly back into the UDF as a string. This CL assumes that the request and response protos would be serialized to / deserialized from a JS string.

Bug: b/321032097
Change-Id: I98b406172ec1b393c4ad9338ae238a30c3bc9389
GitOrigin-RevId: ee8138aca573dbf0cd77e563a0930a9650933ad2