feat: Add test case demonstrating memory functionality
V8 only crashes if config.max_worker_virtual_memory_mb is set too low. Added a test case showing that for appropriately high values, Roma successfully initializes. If the call to set_rlimit_as() is omitted, as was suggested in b/345283494, Executor::Limits::rlimit_as_ defaults to INFINITY, so when this value is set too low, we're constraining our virtual memory too small for V8 to be able to initialize.

Bug: b/345283494
Change-Id: Ib7aba5e0ada5f16b34c18d7b9de9b9c3bca977b0
GitOrigin-RevId: d74f178a43bbb4fa8e0f4787e15f6e5019bf1fd4