fix: Check null when accessing the debuginfo pointer
The ContextImpl is initialized with a default callback that provides
nullptr of debuginfo. This will trigger the server crash when server
does not provide a valid debuginfo but receives a request with consented
debug configuration that sets debug info response to true.

KV server does not support debug info yet.

Bug: b/359892464
Change-Id: I5b62064577ab6026824e594c2ffe239fa8cc9127
GitOrigin-RevId: bbb3f9655e923e0efcbd7938efe5d84486bcb1c1