feat: Support disabling UDF stacktraces in Response
RomaService's Config object has been enhanced with a new field, disable_udf_stacktraces_in_response. When this field is enabled, the message on the absl::StatusOr<ResponseObject> will only contain the error, not the associated stacktrace. This flag does not affect logging functionality, UDF stacktraces will still be logged if a logging function is registered.

Bug: b/352016039
Change-Id: Ie4bcbf3c62802d0eefa129a800bd0d50c4943b4d
GitOrigin-RevId: 6b1139a051625f3b13680611da78059e7e5a3b96