internal: Fail slow for pre-commit runs
With fail fast set to false, it gives the
pre-commit runs a chance to fix all the
issues it can possibly fix automatically
without the need to re-run the pre-commit
over and over again.

This new config has definitely saved
time in B&A dev workflow and can help
the common repo as well.

Bug: NA
Change-Id: Ie3ec45f6949fc899d10731907f16106fb7c0ac82
GitOrigin-RevId: 5ce1f6bc61defefee03822490c5008bc1f6de42b