feat: Copy async_http_client from B&A
Note that AsyncHttpClient is a subset of AsyncClient.
The AsyncClient interface in B&A actually contains
two methods - one for http and one for grpc.
No implementation overrides both, so it doesn't make much
sense to keep both methods in one interface.

Breaking out the http portion makes the clients
pure interfaces.

This is part of a series of CLs to copy
the KV BYOS clients from B&A to common repo.

Bug: 411430242
Change-Id: I1b246bfc43e2e4e1b98ed78ae2b4ff3b4701aea0
GitOrigin-RevId: eb8cc179741f446739a07b057e630b38f81c03ff